Create SPlayerController.CPP
This script gives the player control over his or her character when the game begins, and takes it away when the game ends. The spawning, spectating, and sends messages to the player when he or she is trying to do a forbidden act (picking up another item when his or her inventory is full).